requirements.txt
streamlit
pandas
requests
beautifulsoup4
openpyxl
python-dotenv
